src/main/java/org/jtwig/model/expression/operation/calculators/binary/AndOperationCalculator.java
package org.jtwig.model.expression.operation.calculators.binary;

import org.jtwig.context.RenderContext;
import org.jtwig.model.expression.Expression;
import org.jtwig.model.position.Position;
import org.jtwig.util.JtwigValue;

public class AndOperationCalculator implements BinaryOperationCalculator {
    @Override
    public JtwigValue calculate(RenderContext context, Position position, Expression leftOperand, Expression rightOperand) {
        return new JtwigValue(leftOperand.calculate(context).asBoolean() && rightOperand.calculate(context).asBoolean());
    }
}
